下划线变量改为驼峰形式 数据隔离初步

+++ java/local/src/main/java/cn/toesbieya/jxc/module/request/RateLimiterHelper.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 public class RateLimiterHelper {
     private RateLimiter limiter;
-    private int ipRate;
+    private long ipRate;
 
-    public RateLimiterHelper(int totalRate, int ipRate) {
+    public RateLimiterHelper(long totalRate, long ipRate) {
         this.limiter = RateLimiter.create(totalRate);
         this.ipRate = ipRate;
     }
